@@ -1,10 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const { ScreenshotPageObject } = require('../src/page-objects');
-const { calculateIosTopOffset } = require('../src/page-objects/utils');
-const useBrowser = require('../src/use-browser').default;
+import { ScreenshotPageObject } from '../src/page-objects';
+import { calculateIosTopOffset } from '../src/page-objects/utils';
+import useBrowser from '../src/use-browser';
+import './utils/setup-local-driver';
 
-function setupTest(testFn) {
+type TestFn = (page: ScreenshotPageObject) => Promise<void>;
+function setupTest(testFn: TestFn) {
   return useBrowser(async browser => {
     await browser.url('./test-page-object.html');
     await testFn(new ScreenshotPageObject(browser));
@@ -96,7 +98,7 @@ test(
   'setWindowSize',
   setupTest(async page => {
     await page.setWindowSize({ width: 400, height: 300 });
-    expect(await page.browser.getWindowSize()).toEqual({ width: 400, height: 300 });
+    expect(await page.getViewportSize()).toEqual(expect.objectContaining({ width: 400, height: 300 }));
   })
 );
 
@@ -129,7 +131,7 @@ describe('waitForAssertion', () => {
   test(
     'successful assertion',
     setupTest(async page => {
-      const assertion = jest.fn(() => expect(true).toEqual(true));
+      const assertion = jest.fn(async () => expect(true).toEqual(true));
       await page.waitForAssertion(assertion);
       expect(assertion).toHaveBeenCalledTimes(1);
     })
@@ -139,7 +141,7 @@ describe('waitForAssertion', () => {
     'retrying once assertion',
     setupTest(async page => {
       let counter = 0;
-      const assertion = jest.fn(() => {
+      const assertion = jest.fn(async () => {
         counter++;
         expect(counter).toEqual(2);
       });
@@ -151,7 +153,7 @@ describe('waitForAssertion', () => {
   test(
     'reports the original error into the outer scope',
     setupTest(async page => {
-      const assertion = jest.fn(() => expect(true).toEqual(false));
+      const assertion = jest.fn(async () => expect(true).toEqual(false));
       await expect(page.waitForAssertion(assertion)).rejects.toThrowError(/toEqual/);
       expect(assertion).toHaveBeenCalledTimes(6);
     })
@@ -231,7 +233,7 @@ test(
   setupTest(async page => {
     const height = 250;
     await page.setWindowSize({ width: 300, height });
-    const scrollHeight = await page.getElementProperty('body', 'scrollHeight');
+    const scrollHeight = (await page.getElementProperty('body', 'scrollHeight')) as number;
 
     await page.scrollToBottom('body');
 
