@@ -1,9 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const { promisify } = require('util');
-const useBrowser = require('../src/use-browser').default;
-const { configure } = require('../src/use-browser');
-const { getViewportSize } = require('../src/browser-scripts');
+import { promisify } from 'util';
+import useBrowser from '../src/use-browser';
+import { configure } from '../src/use-browser';
+import { getViewportSize } from '../src/browser-scripts';
+import './utils/setup-local-driver';
+
 const delay = promisify(setTimeout);
 
 test(
@@ -27,7 +29,8 @@ test(
 test('should close browser after test finish', async () => {
   const onDeleteSession = jest.fn();
   await useBrowser(async browser => {
-    browser.overwriteCommand('deleteSession', async originalCommand => {
+    // FIXME: Casting to any isn't ideal, but it's the only way to overwrite this command
+    (browser as any).overwriteCommand('deleteSession', async (originalCommand: any) => {
       await originalCommand();
       onDeleteSession();
     });
@@ -91,7 +94,12 @@ test('should run multiple browsers in parallel', () => {
     await browser.url('./index.html');
     await browser.pause(400);
   });
-  const threadWithDelay = () => delay(200).then(() => useBrowser(browser => browser.url('./index.html'))());
+  const threadWithDelay = () =>
+    delay(200).then(() =>
+      useBrowser(browser => {
+        browser.url('./index.html');
+      })()
+    );
 
   return Promise.all([threadOne(), threadTwo(), threadThree(), threadWithDelay()]);
 });
