@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const { configure } = require('../../src/use-browser');
-const { chromeDriverPort } = require('./config');
+import { chromeDriverPort } from './config';
+import { configure } from '../../src/use-browser';
 
 jest.setTimeout(80 * 1000);
 
