@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 const responses = {
-  pass: () => Promise.resolve({ url: 'http://localhost:4444' }),
+  pass: () => Promise.resolve({ url: 'http://localhost:4444/devicefarm-test-grid' }),
   invalid: () => Promise.resolve({ invalid: true }),
   throttling: () => {
     const error = new Error('ThrottlingException');
